git-svn-id: https://parflow.svn.cloudforge.com/parflow/parflow/trunk@204 39af3fb0-ca4e-0410-adf9-b4ae76d04516

+++ parflow/parflow_lib/well.c
@@ -824,7 +824,7 @@ int           write_header;
 
                fprintf(file, "%2d\n", WellDataPhysicalNumber(well_data_physical));
 
-               fprintf(file, "%d\n", strlen(WellDataPhysicalName(well_data_physical)));
+               fprintf(file, "%d\n", (int)strlen(WellDataPhysicalName(well_data_physical)));
                fprintf(file, "%s\n", WellDataPhysicalName(well_data_physical));
 
                fprintf(file, "%f %f %f %f %f %f %f\n",
@@ -845,7 +845,7 @@ int           write_header;
 
                fprintf(file, "%2d\n", WellDataPhysicalNumber(well_data_physical));
 
-               fprintf(file, "%d\n", strlen(WellDataPhysicalName(well_data_physical)));
+               fprintf(file, "%d\n", (int)strlen(WellDataPhysicalName(well_data_physical)));
                fprintf(file, "%s\n", WellDataPhysicalName(well_data_physical));
 
                fprintf(file, "%f %f %f %f %f %f %f\n",
